User list to file, create User-logic

+++ User.cs
@@ -14,7 +14,7 @@ namespace TheStore
         private string password;
         private string email;
         private string address;
-        private int phone;
+        private double phone;
         private bool loggedIn = false;
 
 
@@ -81,12 +81,12 @@ namespace TheStore
             address = _address;
         }
         
-        public int getPhone()
+        public double getPhone()
         {
             return phone;
         }
 
-        public void setPhone(int _phone) 
+        public void setPhone(double _phone) 
         {
             _phone = phone;
         }
@@ -104,7 +104,7 @@ namespace TheStore
         public User() 
         { 
         }
-        public User(string name, string password, string email, string address, int phone)
+        public User(string name, string password, string email, string address, double phone)
         {
             if (password != null && password.Length > 6)
             {
@@ -119,7 +119,6 @@ namespace TheStore
                 this.phone = phone;
                 UserId = UserId + 1;
                 loggedIn = false;
-                //userList.Add();
         }
        
     }
